+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldAblazeEnchantment.java
@@ -1,14 +1,18 @@
 package com.insane96mcp.shieldsplus.world.item.enchantment;
 
+import com.insane96mcp.shieldsplus.setup.SPEnchantments;
+import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ShieldItem;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 
 public class ShieldAblazeEnchantment extends Enchantment {
 
-	public static final int SECONDS_ON_FIRE = 4;
+	public static final int SECONDS_ON_FIRE = 3;
 
 	public ShieldAblazeEnchantment() {
 		super(Rarity.UNCOMMON, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
@@ -29,5 +33,18 @@ public class ShieldAblazeEnchantment extends Enchantment {
 	@Override
 	public boolean canApplyAtEnchantingTable(ItemStack stack) {
 		return stack.getItem() instanceof ShieldItem;
+	}
+
+	public static void onBlocked(LivingEntity blockingEntity, DamageSource source) {
+		if (!(source.getDirectEntity() instanceof LivingEntity livingEntity))
+			return;
+		apply(blockingEntity, livingEntity);
+	}
+
+	public static void apply(LivingEntity attacker, LivingEntity other) {
+		ItemStack shield = attacker.getUseItem();
+		int ablaze = EnchantmentHelper.getItemEnchantmentLevel(SPEnchantments.ABLAZE.get(), shield);
+		if (ablaze > 0)
+			other.setSecondsOnFire(ablaze * ShieldAblazeEnchantment.SECONDS_ON_FIRE);
 	}
 }