+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldLightweightEnchantment.java
@@ -1,10 +1,16 @@
 package com.insane96mcp.shieldsplus.world.item.enchantment;
 
+import com.insane96mcp.shieldsplus.setup.SPEnchantments;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.ai.attributes.AttributeInstance;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ShieldItem;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 
 import java.util.UUID;
 
@@ -32,5 +38,24 @@ public class ShieldLightweightEnchantment extends Enchantment {
 	@Override
 	public boolean canApplyAtEnchantingTable(ItemStack stack) {
 		return stack.getItem() instanceof ShieldItem;
+	}
+
+	public static void onTick(Player player) {
+		AttributeInstance attribute = player.getAttribute(Attributes.MOVEMENT_SPEED);
+		if (attribute == null)
+			return;
+
+		if (player.isBlocking()) {
+			int lightweight = EnchantmentHelper.getItemEnchantmentLevel(SPEnchantments.LIGHTWEIGHT.get(), player.getUseItem());
+			if (lightweight > 0) {
+
+				if (attribute.getModifier(ShieldLightweightEnchantment.BONUS_SPEED_UUID) == null) {
+					attribute.addTransientModifier(new AttributeModifier(BONUS_SPEED_UUID, "Lightweight bonus speed", BONUS_SPEED * lightweight, AttributeModifier.Operation.MULTIPLY_BASE));
+				}
+			}
+		}
+		else if (attribute.getModifier(ShieldLightweightEnchantment.BONUS_SPEED_UUID) != null) {
+			attribute.removeModifier(ShieldLightweightEnchantment.BONUS_SPEED_UUID);
+		}
 	}
 }